feat(ui): let brush tool handle its events
Move brush tool event logic to its class.
